Edited python files to conform to pep8
Replaced not 'a in i' to 'a not in i'
Replaced 'a == None' and 'a != None to 'a is None' and 'a is not None'
Replaced 'type(x) == type' to 'isinstance(x, type)'
Edited quotes to be consistent.
Replaced 'from x import *' with explicit imports
Used pep8 spacing in function signatures.
Changed classes to inherit from object
Removed unused imports