result_set bugfix
Zero-based boundary condition is not enforced for the variable size_t j,
it was previously free to underflow.

I have placed j-- as a condition within the for loop to enforce this.
See this
thread for considerations on good style for decrementing size_t loops:
http://stackoverflow.com/questions/7224386/iterating-with-size-t-0-as-a-boundary-condition

Placing the j-- at the beginning of the loop is equivalent to --j at the
end, so the functionality is identical.

Updated COPYING as this is standard / mandatory practice for Google employees'
contributions to open source projects.