Added error in load_index python bindings
New error handles the case when a file with an old signature is loaded
Added a test case to test the new error.

Previously C++ printed out an error message and the python side failed
silently. New test cases also handles trying to load an invalid index.
Tests the cases of incorrect file types, different data types, and
different data lengths. Different dimensions does not seem to throw
an error in the C++ code, so that test is commented out.