fix constructor and deconstructor of the node of the hierarchical_clustering_index tree.
This bug will cause the serialization of the hierarchical_clustering_index to crash. Because when saving the index, the pivot and pivot_index for the top node of each tree is not initializaed. then when load these top nodes, pivot = obj->points_[pivot_index]; will cause error that index is out of range.