Normalize line endings for config.h and setup.py.
Currently, CMake configure_file is used to modify files in the source
directory.  Default CMake behavior is to use the host operating system's
line ending.  This can lead to Git deciding that the working tree is dirty.

Here, we standardize line endings for these files using .gitattributes and
use NEWLINE_STYLE CMake option to match with the chosen line endings.